+++ rendu/srcs/ft_strncmp.c
@@ -1,8 +1,8 @@
 #include	"libft.h"
 
-int		ft_strncmp(const char *s1, const char *s2, size_t n)
+int		        ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int     c; 
+	size_t     c; 
 
     c = 0;
     while(s1[c] && s2[c] && c < n)
