+++ rendu/srcs/ft_strnequ.c
@@ -1,8 +1,8 @@
 #include	"libft.h"
 
-int		ft_strnequ(char const *s1,char const *s2, size_t n)
+int			ft_strnequ(char const *s1,char const *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while(s1[i] && s2[i] && i < n)
